Implement new method: it changes the image of toggle switch to on or off
New method implemented - it reads the ImageView clicked by the user, reads the current image
and changes the image to Off switch png if it was originally on, and vice versa.

The images "Toggle_Button_Off.png" and "Toggle_Button_On.png" are imported in to the
resources package in the repository.