Implemented the Start Game method to only start if at least one button is toggled.
Added initialise method that adds all toggle buttons in to an array so they can later be
examined one by one to count the total number of buttons selected.
If there is none, a warning message is displayed. Is there is at least one, it proceeds to the
games screen.